refactor: annotation 패키지로 정리
