+++ .gitignore
@@ -47,3 +47,6 @@ src/main/resources/application.yaml
 
 ### 개발 DB용 docker-compose 제외 ###
 /docker-compose.yaml
+
+# P6Spy log
+spy.log
