+++ src/main/kotlin/com/dh/baro/core/anotation/CheckAuth.kt
@@ -1,4 +1,4 @@
-package com.dh.baro.core.auth
+package com.dh.baro.core.anotation
 
 import com.dh.baro.identity.domain.UserRole
