+++ src/main/kotlin/com/dh/baro/look/domain/Look.kt
@@ -1,7 +1,7 @@
 package com.dh.baro.look.domain
 
 import com.dh.baro.core.AbstractTime
-import com.dh.baro.core.AggregateRoot
+import com.dh.baro.core.anotation.AggregateRoot
 import com.dh.baro.core.IdGenerator
 import jakarta.persistence.*
